+++ extensions/scaleway/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Scaleway Changelog
 
+## [Improvement] - 2022-04-29
+
+- Use the Scaleway JS SDK instead of the REST API
+- Use the `useCachedPromise` hook to fetch data
+
 ## [Initial Version] - 2022-08-05
 
 - Quick access to your serverless containers